Fixed Broken Hours update
I had left some older code in place and the time update was using that old variable when it updated on the hour, so it was reading the same hour digit as when the page was loaded. I moved a block of code that should fix it for now - but I did that from the github site. I'll sit down and make it clean when I get a chance.