Functions tinkering
A couple of small adjustments made to the main function (including re
running the slabtext function when time updates). Prior to the change,
the text would maintain letterspacing and size when the day updated
causing changes (like Friday to Saturday, for instance) to look
horrible. It was a simple oversight when I first put this together.

I've also included the file I use for the demo (index.html) and
improved the style toggle therin.